cleanup of EMTF PH/TH LUTs requested by David L
